+++ tests/functional/adapter/column_tags/fixtures.py
@@ -15,6 +15,7 @@ models:
           pii: "true"
           sensitive: "true"
           key_only: ""
+          null_value:
 """
 
 updated_column_tag_model = """
@@ -32,6 +33,7 @@ models:
           pii: "true"
           sensitive: "true"
           key_only: ""
+          null_value:
 """
 
 column_tags_seed = """